+++ src/components/index.js
@@ -5,6 +5,4 @@ import { fab } from '@fortawesome/free-brands-svg-icons'
 iconLibrary.add(fas, fab)
 
 export { Button } from './buttons/Button/Button.jsx'
-export { PositiveButton } from './buttons/PositiveButton.jsx'
-export { NegativeButton } from './buttons/NegativeButton.jsx'
 export { ThemeProvider } from './contexts/ThemeContext.js'